+++ node/resolvers/search/category.ts
@@ -1,7 +1,7 @@
 import { compose, last, prop, split } from 'ramda'
 
 import { getCategoryInfo, logDegradedSearchError } from './utils'
-import { formatTranslatableProp, shouldTranslateToBinding } from '../../utils/i18n'
+import { shouldTranslateToBinding } from '../../utils/i18n'
 import { Slugify } from '../../utils/slug'
 import { APP_NAME } from './constants'
 
@@ -11,7 +11,7 @@ const lastSegment = compose<string, string[], string>(
 )
 
 function cleanUrl(url: string) {
-  return url.replace(/https:\/\/[A-z0-9]+\.vtexcommercestable\.com\.br/, '').toLowerCase()
+  return url.replace(/https:\/\/[A-Za-z0-9]+\.vtexcommercestable\.com\.br/, '').toLowerCase()
 }
 
 /** This type has to be created because the Catlog API to get category by ID does not return the url or children for now.
@@ -22,10 +22,7 @@ type SafeCategory = CategoryByIdResponse | CategoryTreeResponse
 
 export const resolvers = {
   Category: {
-    name: formatTranslatableProp<SafeCategory, 'name', 'id'>(
-      'name',
-      'id'
-    ),
+    name: prop('name'),
 
     cacheId: prop('id'),
 
@@ -51,15 +48,9 @@ export const resolvers = {
       return settings.slugifyLinks ? Slugify(pathname) : pathname
     },
 
-    metaTagDescription: formatTranslatableProp<SafeCategory, 'MetaTagDescription', 'id'>(
-      'MetaTagDescription',
-      'id'
-    ),
+    metaTagDescription: prop('MetaTagDescription'),
 
-    titleTag: formatTranslatableProp<SafeCategory, 'Title', 'id'>(
-      'Title',
-      'id'
-    ),
+    titleTag: prop('Title'),
 
     slug: async ({ url, id }: SafeCategory, _: unknown, ctx: Context) => {
       if (shouldTranslateToBinding(ctx)) {