+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Use the new client `Catalog-dataplane` that already returns these strings localized
+- Remove `Category` `name`, `titleTag`, and `metaTagDescription` wrapping them for Messages translation.  
+
 ## [1.115.1] - 2026-09-14
 
 ### Changed